SortedList

+++ AmazonMovieReviewer/src/loadata/userMovie.java
@@ -5,7 +5,7 @@ public class userMovie {
 	private String productId;
 	private double rating;
 
-	public userMovie(String productId, int rating) {
+	public userMovie(String productId, double rating) {
 		this.productId = productId;
 		this.rating = rating;
 	}
